Fix null reference exception thrown by RotatePicFrame()
Previously, the script sought to adjust the rotation of the current polaroid, before the current polaroid was even set.
The orientation value is now provided by PhotoAccessManager.cs and passed to PolaroidManager ref; once a
polaroid is spawned, the stored texture and stored orientation value are used to adjust the PicFrame appropriately.